[Disk Manager]: Remove redundant type conversion (#2075)

+++ cloud/disk_manager/internal/pkg/dataplane/replicate_disk_task.go
@@ -134,8 +134,8 @@ func (t *replicateDiskTask) saveProgress(
 		return err
 	}
 
-	bytesPerSecond := uint64(performance.ConvertMiBsToBytes(
-		t.performanceConfig.GetReplicateDiskBandwidthMiBs()),
+	bytesPerSecond := performance.ConvertMiBsToBytes(
+		t.performanceConfig.GetReplicateDiskBandwidthMiBs(),
 	)
 
 	if t.state.ChunkCount != 0 && t.state.Progress != 1 {
